# generated-model-resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Denaly | API",
    "description": "Swagger UI for Denaly API",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "API Support",
      "url": "https://www.example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:4000/web",
      "description": "Local server"
    },
    {
      "url": "https://development.gigantic-server.com",
      "description": "Development server"
    },
    {
      "url": "https://staging.gigantic-server.com",
      "description": "Staging server"
    },
    {
      "url": "https://api.gigantic-server.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/s3-objects": {
      "get": {
        "tags": [
          "S3Object"
        ],
        "summary": "Get all S3Objects",
        "description": "Get all S3Objects",
        "operationId": "getS3Objects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of S3Objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/S3ObjectEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "S3Object"
        ],
        "summary": "Create a new S3Object",
        "description": "Create a new S3Object",
        "operationId": "createS3Object",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateS3ObjectDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "model_type": "",
                "model_id": 1,
                "mimetype": "",
                "is_private": true,
                "region": "",
                "bucket": "",
                "key": ""
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created S3Object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3ObjectEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/s3-objects/{id}": {
      "get": {
        "tags": [
          "S3Object"
        ],
        "summary": "Get S3Object by ID",
        "description": "Get S3Object by ID",
        "operationId": "getS3ObjectById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "S3Object found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3ObjectEntity"
                }
              }
            }
          },
          "404": {
            "description": "S3object not found"
          }
        }
      },
      "put": {
        "tags": [
          "S3Object"
        ],
        "summary": "Update S3Object by ID",
        "description": "Update S3Object by ID",
        "operationId": "updateS3Object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateS3ObjectDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "model_type": "",
                "model_id": 1,
                "mimetype": "",
                "is_private": true,
                "region": "",
                "bucket": "",
                "key": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "S3Object updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3ObjectEntity"
                }
              }
            }
          },
          "404": {
            "description": "S3Object not found"
          }
        }
      },
      "patch": {
        "tags": [
          "S3Object"
        ],
        "summary": "Patch S3Object by ID",
        "description": "Patch S3Object by ID",
        "operationId": "PatchS3Object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateS3ObjectDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "model_type": "",
                "model_id": 1,
                "mimetype": "",
                "is_private": true,
                "region": "",
                "bucket": "",
                "key": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "S3Object updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3ObjectEntity"
                }
              }
            }
          },
          "404": {
            "description": "S3Object not found"
          }
        }
      },
      "delete": {
        "tags": [
          "S3Object"
        ],
        "summary": "Delete S3Object by ID",
        "description": "Delete S3Object by ID",
        "operationId": "deleteS3Object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "S3Object deleted"
          },
          "404": {
            "description": "S3Object not found"
          }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Get all Tags",
        "description": "Get all Tags",
        "operationId": "getTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of Tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tag"
        ],
        "summary": "Create a new Tag",
        "description": "Create a new Tag",
        "operationId": "createTag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "name": "",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Tag",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/tags/{id}": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Get Tag by ID",
        "description": "Get Tag by ID",
        "operationId": "getTagById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagEntity"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found"
          }
        }
      },
      "put": {
        "tags": [
          "Tag"
        ],
        "summary": "Update Tag by ID",
        "description": "Update Tag by ID",
        "operationId": "updateTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "name": "",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagEntity"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Tag"
        ],
        "summary": "Patch Tag by ID",
        "description": "Patch Tag by ID",
        "operationId": "PatchTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "name": "",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagEntity"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Tag"
        ],
        "summary": "Delete Tag by ID",
        "description": "Delete Tag by ID",
        "operationId": "deleteTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tag deleted"
          },
          "404": {
            "description": "Tag not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "S3ObjectEntity": {
        "type": "object",
        "description": "S3Object entity",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "model_type": {
            "type": "string",
            "format": "string",
            "description": "Model type"
          },
          "model_id": {
            "type": "integer",
            "format": "int64",
            "description": "Model id"
          },
          "mimetype": {
            "type": "string",
            "format": "string",
            "description": "Mimetype"
          },
          "is_private": {
            "type": "boolean",
            "format": "boolean",
            "description": "Is private"
          },
          "region": {
            "type": "string",
            "format": "string",
            "description": "Region"
          },
          "bucket": {
            "type": "string",
            "format": "string",
            "description": "Bucket"
          },
          "key": {
            "type": "string",
            "format": "string",
            "description": "Key"
          }
        },
        "required": [
          "id",
          "create_at",
          "is_private",
          "region",
          "bucket",
          "key"
        ]
      },
      "CreateS3ObjectDto": {
        "type": "object",
        "description": "Create S3Object DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "model_type": {
            "type": "string",
            "format": "string",
            "description": "Model type"
          },
          "model_id": {
            "type": "integer",
            "format": "int64",
            "description": "Model id"
          },
          "mimetype": {
            "type": "string",
            "format": "string",
            "description": "Mimetype"
          },
          "is_private": {
            "type": "boolean",
            "format": "boolean",
            "description": "Is private"
          },
          "region": {
            "type": "string",
            "format": "string",
            "description": "Region"
          },
          "bucket": {
            "type": "string",
            "format": "string",
            "description": "Bucket"
          },
          "key": {
            "type": "string",
            "format": "string",
            "description": "Key"
          }
        }
      },
      "UpdateS3ObjectDto": {
        "type": "object",
        "description": "Update S3Object DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "model_type": {
            "type": "string",
            "format": "string",
            "description": "Model type"
          },
          "model_id": {
            "type": "integer",
            "format": "int64",
            "description": "Model id"
          },
          "mimetype": {
            "type": "string",
            "format": "string",
            "description": "Mimetype"
          },
          "is_private": {
            "type": "boolean",
            "format": "boolean",
            "description": "Is private"
          },
          "region": {
            "type": "string",
            "format": "string",
            "description": "Region"
          },
          "bucket": {
            "type": "string",
            "format": "string",
            "description": "Bucket"
          },
          "key": {
            "type": "string",
            "format": "string",
            "description": "Key"
          }
        }
      },
...
      "TagEntity": {
        "type": "object",
        "description": "Tag entity",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "name": {
            "type": "string",
            "format": "string",
            "description": "Name"
          },
          "description": {
            "type": "string",
            "format": "string",
            "description": "Description"
          }
        },
        "required": [
          "id",
          "create_at",
          "name"
        ]
      },
      "CreateTagDto": {
        "type": "object",
        "description": "Create Tag DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "name": {
            "type": "string",
            "format": "string",
            "description": "Name"
          },
          "description": {
            "type": "string",
            "format": "string",
            "description": "Description"
          }
        }
      },
      "UpdateTagDto": {
        "type": "object",
        "description": "Update Tag DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "name": {
            "type": "string",
            "format": "string",
            "description": "Name"
          },
          "description": {
            "type": "string",
            "format": "string",
            "description": "Description"
          }
        }
      },
...
    }
  }
}